The beginning of the finishing touches
Refactored, removed unused imports and updated the views.
