+++ src/myattendance/GUI/Controller/TeacherAttendanceOverviewController.java
@@ -1,26 +1,18 @@
 package myattendance.GUI.Controller;
 
 import com.sun.javafx.scene.control.skin.DatePickerSkin;
-import com.sun.prism.paint.Color;
-import java.awt.event.ActionListener;
 
 import java.io.IOException;
 import java.net.URL;
-import java.time.DayOfWeek;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.ZoneId;
-import java.util.ArrayList;
 import java.util.Date;
-import java.util.List;
 import java.util.Optional;
 import java.util.ResourceBundle;
 import java.util.Timer;
 import java.util.TimerTask;
 import javafx.application.Platform;
-import javafx.beans.property.DoubleProperty;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
@@ -37,24 +29,18 @@ import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.ComboBox;
-import javafx.scene.control.DateCell;
 import javafx.scene.control.DatePicker;
 import javafx.scene.control.Label;
-import javafx.scene.control.MenuBar;
 import javafx.scene.control.Pagination;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
-import javafx.scene.control.Tooltip;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.Region;
 import javafx.scene.layout.VBox;
-import static javafx.scene.paint.Color.GREEN;
-import javafx.scene.paint.Paint;
 import javafx.stage.Stage;
-import javafx.stage.Window;
 import javafx.util.Callback;
 import myattendance.BE.Course;
 import myattendance.BE.Day;
@@ -81,7 +67,6 @@ public class TeacherAttendanceOverviewController implements Initializable
     User lastSelectedStudent;
     Day clickedDay;
 
-    List<Day> nonSchoolDays = new ArrayList<>();
 
     Course lastSelectedCourse;
 
@@ -146,7 +131,6 @@ public class TeacherAttendanceOverviewController implements Initializable
     @Override
     public void initialize(URL url, ResourceBundle rb)
     {
-        nonSchoolDays = dateParser.listNonSchoolDays();
         showConstantCalender();
         setClickCal();
         updatePresentCounter();
@@ -208,10 +192,7 @@ public class TeacherAttendanceOverviewController implements Initializable
     {
         calendar = new DatePicker(LocalDate.now());
 
-        // Factory to create Cell of DatePicker
-        Callback<DatePicker, DateCell> dayCellFactory = this.getDayCellFactory();
-
-        calendar.setDayCellFactory(getDayCellFactory());
+        calendar.setDayCellFactory(dateParser.getDayCellFactory());
         DatePickerSkin datePickerSkin = new DatePickerSkin(calendar);
         Region pop = (Region) datePickerSkin.getPopupContent();
 
@@ -219,42 +200,6 @@ public class TeacherAttendanceOverviewController implements Initializable
 
         vBoxSelectionContent.setSpacing(200);
         vBoxSelectionContent.getChildren().add(pop);
-    }
-
-    // Factory to create Cell of DatePicker
-    private Callback<DatePicker, DateCell> getDayCellFactory()
-    {
-
-        final Callback<DatePicker, DateCell> dayCellFactory = new Callback<DatePicker, DateCell>()
-        {
-
-            @Override
-            public DateCell call(final DatePicker datePicker)
-            {
-                return new DateCell()
-                {
-                    @Override
-                    public void updateItem(LocalDate item, boolean empty)
-                    {
-
-                        super.updateItem(item, empty);
-
-                        //dateParser.checkNonSchoolDay(item);
-                        //Instant instant = Instant.from(item);
-                        //DateTime date = new DateTime(item.atStartOfDay());
-                        for (Day day : nonSchoolDays)
-                        {
-                            if (day.getDateInTime().toLocalDate().toString().equals(item.toString()))
-                            {
-                                setTooltip(new Tooltip("Not school day"));
-                                setStyle("-fx-background-color: #C0C0C0;");
-                            }
-                        }
-                    }
-                };
-            }
-        };
-        return dayCellFactory;
     }
 
     private void setClickCal()
@@ -427,7 +372,7 @@ public class TeacherAttendanceOverviewController implements Initializable
     private void automaticUpdate()
     {
         // The time between every update in milliseconds
-        int delay = 150000; // REMEMBER TO REMOVE A 0 SO IT'S 15000
+        int delay = 1200000; // 120 seconds
 
 
         // Creates a new timer
