+++ src/myattendance/GUI/Model/DateParser.java
@@ -1,12 +1,15 @@
 package myattendance.GUI.Model;
 
-import java.util.Date;
 import java.util.List;
 import myattendance.BE.Day;
 import myattendance.BE.User;
 import myattendance.BLL.BLLFacade;
 import org.joda.time.DateTime;
-import java.time.Instant;
+import java.time.LocalDate;
+import javafx.scene.control.DateCell;
+import javafx.scene.control.DatePicker;
+import javafx.scene.control.Tooltip;
+import javafx.util.Callback;
 
 /**
  *
@@ -56,6 +59,42 @@ public class DateParser
     public List<Day> listNonSchoolDays()
     {
         return bllFacade.listNonSchoolDays();
+    }
+
+    // Factory to create Cell of DatePicker
+    public Callback<DatePicker, DateCell> getDayCellFactory()
+    {
+
+        final Callback<DatePicker, DateCell> dayCellFactory = new Callback<DatePicker, DateCell>()
+        {
+
+            @Override
+            public DateCell call(final DatePicker datePicker)
+            {
+                return new DateCell()
+                {
+                    @Override
+                    public void updateItem(LocalDate item, boolean empty)
+                    {
+
+                        super.updateItem(item, empty);
+
+                        //dateParser.checkNonSchoolDay(item);
+                        //Instant instant = Instant.from(item);
+                        //DateTime date = new DateTime(item.atStartOfDay());
+                        for (Day day : listNonSchoolDays())
+                        {
+                            if (day.getDateInTime().toLocalDate().toString().equals(item.toString()))
+                            {
+                                setTooltip(new Tooltip("Not school day"));
+                                setStyle("-fx-background-color: #C0C0C0;");
+                            }
+                        }
+                    }
+                };
+            }
+        };
+        return dayCellFactory;
     }
 
 }
