+++ src/myattendance/GUI/Model/StudentViewModel.java
@@ -1,16 +1,12 @@
 package myattendance.GUI.Model;
 
-import java.text.DecimalFormat;
-import java.util.ArrayList;
 import java.util.List;
-import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.scene.chart.PieChart;
 import javafx.scene.chart.XYChart;
 import myattendance.BE.Day;
 import myattendance.BE.User;
 import myattendance.BLL.BLLFacade;
-import org.joda.time.DateTime;
 
 /**
  *
